Fix type hint: interval_seconds should be float, not int
Changed interval_seconds type hint from int to float in _heartbeat_keeper
to match the signature of threading.Event.wait(timeout=...) which expects
a float.

This fixes the Sonar type violation where the test passes 0.001 (a float)
for fast execution. The change is backward compatible since:
- int is a subtype of float in Python's type system
- All existing production code passes int values which work fine
- threading.Event.wait() accepts both int and float

Benefits:
- Aligns type hint with actual API contract (Event.wait expects float)
- Enables sub-second precision for tests (and production if needed)
- No behavior change - all existing code continues to work

Agent-Logs-Url: https://github.com/DashFin-FarDb/financial-asset-relationship-db/sessions/bf6d4db7-d8cc-4d20-a810-3f781f1dfc90

+++ api/routers/graph_admin.py
@@ -694,7 +694,7 @@ def _heartbeat_keeper(
     worker_id: str,
     stop_event: threading.Event,
     lock_lost_event: threading.Event,
-    interval_seconds: int,
+    interval_seconds: float,
 ) -> None:
     """Background thread that periodically refreshes both the distributed lock and rebuild heartbeat."""
     while not stop_event.wait(timeout=interval_seconds):
